[ELI-67] Apply suggestions from code review
