+++ core/src/test/java/com/exadel/etoolbox/linkinspector/core/services/mocks/MockDataFeedService.java
@@ -1,6 +1,6 @@
 package com.exadel.etoolbox.linkinspector.core.services.mocks;
 
-import com.exadel.etoolbox.linkinspector.core.models.UpdatedItem;
+import com.exadel.etoolbox.linkinspector.core.services.data.models.UpdatedItem;
 import com.exadel.etoolbox.linkinspector.core.services.data.DataFeedService;
 import com.exadel.etoolbox.linkinspector.core.services.data.models.DataFilter;
 import com.exadel.etoolbox.linkinspector.core.services.data.models.GridResource;
